REFACTORING: folder structure and React files extensions updated
