+++ src/__tests__/App.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from '../components/App/App.jsx';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
