+++ src/components/Viewpoint/Viewpoint.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import by from 'sort-by';
-import Topic from './Topic';
+import Topic from '../Topic/Topic.jsx';
 
 class Viewpoint extends Component {
   render() {
